Refactor statikvariable functions into twig extensions so statikvariable is clean for custom project code
