+++ templates/jsPlugins/ajaxPaging.twig
@@ -39,7 +39,7 @@
                 {% endfor %}
             </div>
             <div class="flex justify-center mt-10 js-ajax-paging-links">
-                {{ craft.statik.paginate(pageInfoNews) | raw }}
+                {{ render_pagination(pageInfoNews) | raw }}
             </div>
         </div>
     </div>
@@ -60,7 +60,7 @@
                 {% endfor %}
             </div>
             <div class="flex justify-center mt-10 js-ajax-paging-links">
-                {{ craft.statik.paginate(pageInfoNews2) | raw }}
+                {{ render_pagination(pageInfoNews2) | raw }}
             </div>
         </div>
     </div>
